requirements.txt
# Web framework
fastapi==0.135.1
uvicorn==0.41.0
starlette==0.52.1

# DB
SQLAlchemy==2.0.48
asyncpg
psycopg2-binary==2.9.11
alembic==1.18.4

# Auth / Security
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.5.0

# Pydantic / Settings
pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.3.0

# Redis
redis==7.3.0

# Email
fastapi-mail==1.6.2
aiosmtplib==5.1.0

# OAuth (HTTP 요청)
requests==2.32.5

# Utils
python-dotenv==1.2.2
python-multipart
